chore(graph-gateway): rename network module entities to snapshot
