+++ graph-gateway/src/network.rs
@@ -2,15 +2,14 @@
 //! provides information about the subgraphs (and subgraph deployments) registered in the network
 //! smart contract, as well as the indexers that are indexing them.
 
-pub use entities::{
-    BlockNumber, DeploymentId, GraphNetwork, Indexer, Indexing, IndexingId, IndexingStatus,
-    SubgraphId,
-};
 pub use service::{
     NetworkService, NetworkServiceBuilder, NetworkServicePending, ResolvedSubgraphInfo,
 };
+pub use snapshot::{
+    BlockNumber, DeploymentId, Indexer, Indexing, IndexingId, IndexingStatus,
+    NetworkTopologySnapshot, SubgraphId,
+};
 
-mod entities;
 pub mod indexers_addr_blocklist;
 pub mod indexers_cost_model_compiler;
 pub mod indexers_cost_model_resolver;
@@ -21,4 +20,5 @@ pub mod indexers_poi_blocklist;
 pub mod indexers_poi_resolver;
 pub mod internal;
 mod service;
+mod snapshot;
 pub mod subgraph;
